+++ src/tnt.cpp
@@ -546,22 +546,22 @@ namespace TA3D
                 float dh=0.0f;
                 if (y > 0)
                 {
-                    float dz = fabs( map->get_zdec( x, y) - map->get_zdec( x, y - 1) + 8.0f);
+                    float dz = fabsf( map->get_zdec( x, y) - map->get_zdec( x, y - 1) + 8.0f);
                     if (dz == 0.0f)	dz = 100000000.0f;
                     else			dz = 8.0f / dz;
-                    dh = Math::Max(dh,(float)fabs(map->h_map[y][x]-map->h_map[y-1][x]) * dz);
+                    dh = Math::Max(dh,(float)fabsf(map->h_map[y][x]-map->h_map[y-1][x]) * dz);
                 }
                 if (y + 1 < (map->bloc_h << 1))
                 {
-                    float dz = fabs( map->get_zdec( x, y + 1) - map->get_zdec( x, y) + 8.0f);
+                    float dz = fabsf( map->get_zdec( x, y + 1) - map->get_zdec( x, y) + 8.0f);
                     if (dz == 0.0f)	dz = 100000000.0f;
                     else				dz = 8.0f / dz;
-                    dh = Math::Max(dh,(float)fabs(map->h_map[y][x]-map->h_map[y+1][x]) * dz);
+                    dh = Math::Max(dh,(float)fabsf(map->h_map[y][x]-map->h_map[y+1][x]) * dz);
                 }
                 if (x > 0)
-                    dh = Math::Max(dh,(float)fabs(map->h_map[y][x]-map->h_map[y][x-1]));
+                    dh = Math::Max(dh,(float)fabsf(map->h_map[y][x]-map->h_map[y][x-1]));
                 if (x + 1 < (map->bloc_w << 1))
-                    dh = Math::Max(dh,(float)fabs(map->h_map[y][x]-map->h_map[y][x+1]));
+                    dh = Math::Max(dh,(float)fabsf(map->h_map[y][x]-map->h_map[y][x+1]));
                 map->map_data[y][x].dh=dh;
             }
         }
